+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## March 16, 2021
+
+- Release v0.3.
+- Widgets Added: ExchangeRate
+- Widgets Updated: Calendar (styles)
+
 ## March 2, 2021
 
 - Release v0.2.